rename plugins package to electrum_plugins
The advantages are:

* local and installed packages are imported the same way
* unit tests do test modified code in local repo
* absolute imports instead of relative imports can be used
