+++ contrib/build-osx/osx.spec
@@ -31,7 +31,7 @@ datas = [
     (electrum+'lib/*.json', PYPKG),
     (electrum+'lib/wordlist/english.txt', PYPKG + '/wordlist'),
     (electrum+'lib/locale', PYPKG + '/locale'),
-    (electrum+'plugins', PYPKG + '_plugins'),
+    (electrum + 'electrum_plugins', 'electrum_plugins'),
 ]
 datas += collect_data_files('trezorlib')
 datas += collect_data_files('btchip')
@@ -54,12 +54,12 @@ a = Analysis([electrum+MAIN_SCRIPT,
               electrum+'lib/bitcoin.py',
               electrum+'lib/dnssec.py',
               electrum+'lib/commands.py',
-              electrum+'plugins/cosigner_pool/qt.py',
-              electrum+'plugins/email_requests/qt.py',
-              electrum+'plugins/trezor/client.py',
-              electrum+'plugins/trezor/qt.py',
-              electrum+'plugins/keepkey/qt.py',
-              electrum+'plugins/ledger/qt.py',
+              electrum+'electrum_plugins/cosigner_pool/qt.py',
+              electrum+'electrum_plugins/email_requests/qt.py',
+              electrum+'electrum_plugins/trezor/client.py',
+              electrum+'electrum_plugins/trezor/qt.py',
+              electrum+'electrum_plugins/keepkey/qt.py',
+              electrum+'electrum_plugins/ledger/qt.py',
               ],
              binaries=binaries,
              datas=datas,
