+++ electrum_plugins/README
