+++ electrum_plugins/cosigner_pool/qt.py
