+++ electrum_plugins/hw_wallet/cmdline.py
