+++ electrum_plugins/keepkey/clientbase.py
