+++ electrum_plugins/keepkey/cmdline.py
