+++ electrum_plugins/keepkey/qt.py
