+++ electrum_plugins/labels/cmdline.py
