+++ electrum_plugins/labels/kivy.py
