+++ electrum_plugins/labels/labels.py
