+++ electrum_plugins/labels/qt.py
