+++ electrum_plugins/revealer/DejaVuSansMono-Bold.ttf
