+++ electrum_plugins/revealer/LICENSE_DEJAVU.txt
