+++ electrum_plugins/revealer/SIL Open Font License.txt
