+++ electrum_plugins/revealer/SourceSansPro-Bold.otf
