+++ electrum_plugins/revealer/qt.py
