+++ electrum_plugins/trezor/client.py
