+++ electrum_plugins/trezor/clientbase.py
