+++ electrum_plugins/trezor/cmdline.py
