+++ electrum_plugins/trezor/qt.py
