+++ electrum_plugins/trezor/transport.py
