+++ electrum_plugins/trezor/trezor.py
