+++ electrum_plugins/trustedcoin/__init__.py
