+++ electrum_plugins/trustedcoin/cmdline.py
