+++ electrum_plugins/trustedcoin/kivy.py
