+++ electrum_plugins/trustedcoin/qt.py
